% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osrm.backend-package.R
\docType{package}
\name{osrm.backend-package}
\alias{osrm.backend}
\alias{osrm.backend-package}
\title{osrm.backend: Bindings for 'Open Source Routing Machine'}
\description{
Install and control 'Open Source Routing Machine' ('OSRM') backend executables to prepare routing data and run/stop a local 'OSRM' server. For computations with the running server use the 'osrm' R package (\url{https://cran.r-project.org/package=osrm}).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/e-kotov/osrm.backend}
  \item \url{https://www.ekotov.pro/osrm.backend/}
  \item Report bugs at \url{https://github.com/e-kotov/osrm.backend/issues}
}

}
\author{
\strong{Maintainer}: Egor Kotov \email{kotov.egor@gmail.com} (\href{https://orcid.org/0000-0001-6690-5345}{ORCID}) [copyright holder]

}
\keyword{internal}
