% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osrm_stop_server.R
\name{osrm_stop_all}
\alias{osrm_stop_all}
\title{Stop all running OSRM servers started via this package}
\usage{
osrm_stop_all()
}
\value{
The number of servers attempted.
}
\description{
Stop all running OSRM servers started via this package
}
\examples{
\dontrun{
# Stop any registered servers, regardless of how they were started
stopped <- osrm_stop_all()
stopped
}
}
