% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osrm_stop_server.R
\name{osrm_stop_server}
\alias{osrm_stop_server}
\title{Stop an OSRM server started with \code{osrm_start_server}}
\usage{
osrm_stop_server(server)
}
\arguments{
\item{server}{A \code{processx::process} object returned by \code{\link[=osrm_start_server]{osrm_start_server()}}.}
}
\value{
Invisibly returns the same \code{processx::process} object (now stopped).
}
\description{
Terminates an \code{osrm-routed} process that was launched by \code{\link[=osrm_start_server]{osrm_start_server()}}.
}
